Fixed class structure and removed duplicated classes. Slightly adjusted Student creation to have majors and minors as well as a username and password.
