+++ src/test/java/ithacacollege/comp345/group4/classPlanner/StudentTest.java
@@ -1,6 +1,7 @@
 package ithacacollege.comp345.group4.classPlanner;
 
 import ithacacollege.comp345.group4.classPlanner.model.Course;
+import ithacacollege.comp345.group4.classPlanner.model.Student;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
@@ -12,7 +13,7 @@ public class StudentTest {
 
     @Test
     void addPastCoursesTest(){
-        Student student = new Student(12345678, null, null);
+        Student student = new Student("test", "abc", null, null);
         List<Course> courseList = new ArrayList<>();
         Course course1 = new Course("test1", 12345, 3.0, "TEST123", "FA17", null);
         Course course2 = new Course("test2", 13465, 3.0, "TEST321", "FA17", null);
@@ -38,7 +39,7 @@ public class StudentTest {
 
     @Test
     void addCurrentCoursesTest(){
-        Student student = new Student(12345678, null, null);
+        Student student = new Student("test", "abc", null, null);
         List<Course> courseList = new ArrayList<>();
         Course course1 = new Course("Software Engineering", 12345, 3.0, "COMP345", "FA17", null);
         Course course2 = new Course("Machine Learning", 13465, 3.0, "COMP490", "FA17", null);
@@ -63,7 +64,7 @@ public class StudentTest {
 
     @Test
     void addPlannedCoursesTest(){
-        Student student = new Student(12345678, null, null);
+        Student student = new Student("test", "abc", null, null);
         List<Course> courseList = new ArrayList<>();
         Course course1 = new Course("test1", 12345, 3.0, "TEST123", "FA17", null);
         Course course2 = new Course("test2", 13465, 3.0, "TEST321", "FA17", null);
@@ -88,7 +89,7 @@ public class StudentTest {
 
     @Test
     void viewCoursesTest(){
-        Student student = new Student(12345678, null, null);
+        Student student = new Student("test", "abc", null, null);
         List<Course> courseList = new ArrayList<>();
         Course course1 = new Course("Software Engineering", 12345, 3.0, "COMP345", "FA17", null);
         Course course2 = new Course("Machine Learning", 13465, 3.0, "COMP490", "FA17", null);
